refactor: 프로젝트에 맞게 프로젝트 구조 변경
Konnect로 이름 변경 및 구조 변경
