+++ src/main/java/com/konnect/KonnectApplication.java
@@ -1,0 +1,13 @@
+package com.konnect;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class KonnectApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(KonnectApplication.class, args);
+    }
+
+}
