+++ src/main/java/com/konnect/controller/MainController.java
@@ -1,0 +1,15 @@
+package com.konnect.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+@Controller
+public class MainController {
+
+    @GetMapping("/")
+    @ResponseBody
+    public String mainAPI() {
+        return "main route";
+    }
+}
