+++ src/main/java/com/konnect/dto/KakaoResponse.java
@@ -1,0 +1,41 @@
+package com.konnect.dto;
+
+import lombok.extern.slf4j.Slf4j;
+
+import java.util.Map;
+
+@Slf4j
+public class KakaoResponse implements OAuth2Response {
+    private final Map<String, Object> attribute;
+    private final Map<String, Object> kakaoAccount;
+    private final Map<String, Object> profile;
+
+    public KakaoResponse(Map<String, Object> attribute) {
+        this.attribute = attribute;
+        this.kakaoAccount = (Map<String, Object>) attribute.get("kakao_account");
+        this.profile = (Map<String, Object>) kakaoAccount.get("profile");
+        log.debug("Kakao attribute info: {}", attribute);
+        log.debug("Kakao account info: {}", kakaoAccount);
+        log.debug("Kakao profile info: {}", profile);
+    }
+
+    @Override
+    public String getProvider() {
+        return "kakao";
+    }
+
+    @Override
+    public String getProviderId() {
+        return attribute.get("id").toString();
+    }
+
+    @Override
+    public String getEmail() {
+        return kakaoAccount.get("email").toString();
+    }
+
+    @Override
+    public String getName() {
+        return profile.get("nickname").toString();
+    }
+}
