+++ src/main/java/com/konnect/repository/UserRepository.java
@@ -1,0 +1,8 @@
+package com.konnect.repository;
+
+import com.konnect.entity.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
+    UserEntity findByOauthCode(String oauthCode);
+}
