--- src/main/java/oauth/test/oauth2jwt/OAuth2JwtApplication.java
@@ -1,13 +1,0 @@
-package oauth.test.oauth2jwt;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OAuth2JwtApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(OAuth2JwtApplication.class, args);
-    }
-
-}
