--- src/main/java/oauth/test/oauth2jwt/config/CorsMvcConfig.java
@@ -1,20 +1,0 @@
-package oauth.test.oauth2jwt.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class CorsMvcConfig implements WebMvcConfigurer {
-
-    @Value("${client.url}")
-    String clientUrl;
-
-    @Override
-    public void addCorsMappings(CorsRegistry corsRegistry) {
-        corsRegistry.addMapping("/**")
-                .exposedHeaders("Set-Cookie")
-                .allowedOrigins(clientUrl);
-    }
-}
