--- src/main/java/oauth/test/oauth2jwt/config/SecurityConfig.java
@@ -1,119 +1,0 @@
-package oauth.test.oauth2jwt.config;
-
-
-import jakarta.servlet.http.HttpServletRequest;
-import lombok.RequiredArgsConstructor;
-import oauth.test.oauth2jwt.jwt.CustomAuthenticationEntryPoint;
-import oauth.test.oauth2jwt.jwt.JWTFilter;
-import oauth.test.oauth2jwt.jwt.JWTUtil;
-import oauth.test.oauth2jwt.oauth2.CustomSuccessHandler;
-import oauth.test.oauth2jwt.service.CustomOAuth2UserService;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.env.Environment;
-import org.springframework.core.env.Profiles;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.oauth2.client.web.OAuth2LoginAuthenticationFilter;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.CorsConfigurationSource;
-
-import java.util.Collections;
-
-@Configuration
-@EnableWebSecurity
-@RequiredArgsConstructor
-public class SecurityConfig {
-
-    private final Environment env;
-    private final CustomOAuth2UserService customOAuth2UserService;
-    private final CustomSuccessHandler customSuccessHandler;
-    private final JWTUtil jwtUtil;
-    private final CustomAuthenticationEntryPoint customAuthenticationEntryPoint;
-
-    @Value("${client.url}")
-    String clientUrl;
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http
-                .cors(corsCustomizer -> corsCustomizer.configurationSource(request -> {
-                    CorsConfiguration configuration = new CorsConfiguration();
-
-                    configuration.setAllowedOrigins(Collections.singletonList(clientUrl));
-                    configuration.setAllowedMethods(Collections.singletonList("*"));
-                    configuration.setAllowCredentials(true);
-                    configuration.setAllowedHeaders(Collections.singletonList("*"));
-                    configuration.setMaxAge(3600L);
-
-                    configuration.setExposedHeaders(Collections.singletonList("Set-Cookie"));
-                    configuration.setExposedHeaders(Collections.singletonList("Authorization"));
-
-                    return configuration;
-                }));
-        
-        // 인증 오류 핸들링 추가(JWT 만료)
-        http
-                .exceptionHandling(ex -> ex
-                        .authenticationEntryPoint(customAuthenticationEntryPoint)
-                );
-
-        //csrf disable(어차피 JWT 사용해서 stateless 상태로 관리할 것이므로 필요 X)
-        http
-                .csrf((auth) -> auth.disable());
-
-        //Form 로그인 방식 disable
-        http
-                .formLogin((auth) -> auth.disable());
-
-        //HTTP Basic 인증 방식 disable
-        http
-                .httpBasic((auth) -> auth.disable());
-
-        //JWTFilter 추가
-        // (위치는 OAuth2LoginAuthenticationFilter 이후 등록)
-        http
-                .addFilterAfter(new JWTFilter(jwtUtil), OAuth2LoginAuthenticationFilter.class);
-
-
-        //oauth2 -> oauth2 를 통해, 인증 진행
-        http
-//                .oauth2Login(Customizer.withDefaults());
-                .oauth2Login((oauth2) -> oauth2
-                        .userInfoEndpoint(userInfoEndpointConfig -> userInfoEndpointConfig
-                                .userService(customOAuth2UserService))
-                        .successHandler(customSuccessHandler)
-                );
-        //경로별 인가 작업
-        http
-                .authorizeHttpRequests(auth -> {
-                    // 공통 허용 경로
-                    auth.requestMatchers("/").permitAll();
-
-                    // dev 환경일 때, Swagger 경로 허용
-                    if (env.acceptsProfiles(Profiles.of("dev"))) {
-                        auth.requestMatchers(
-                                "/swagger-ui.html",
-                                "/swagger-ui/**",
-                                "/v3/api-docs/**",
-                                "/swagger-resources/**",
-                                "/webjars/**"
-                        ).permitAll();
-                    }
-
-                    // 나머지는 인증 필요
-                    auth.anyRequest().authenticated();
-                });
-
-        //세션 설정 : STATELESS
-        http
-                .sessionManagement((session) -> session
-                        .sessionCreationPolicy(SessionCreationPolicy.STATELESS));
-
-        return http.build();
-    }
-}
