--- src/main/java/oauth/test/oauth2jwt/controller/MainController.java
@@ -1,15 +1,0 @@
-package oauth.test.oauth2jwt.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-@Controller
-public class MainController {
-
-    @GetMapping("/")
-    @ResponseBody
-    public String mainAPI() {
-        return "main route";
-    }
-}
