--- src/main/java/oauth/test/oauth2jwt/dto/CustomOAuth2User.java
@@ -1,43 +1,0 @@
-package oauth.test.oauth2jwt.dto;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Map;
-
-@RequiredArgsConstructor
-public class CustomOAuth2User implements OAuth2User {
-
-    private final UserDTO userDTO;
-
-    @Override
-    public Map<String, Object> getAttributes() {
-        return Map.of();
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        Collection<GrantedAuthority> collection = new ArrayList<>();
-
-        collection.add(new GrantedAuthority() {
-            @Override
-            public String getAuthority() {
-                return userDTO.getRole();
-            }
-        });
-
-        return collection;
-    }
-
-    @Override
-    public String getName() {
-        return userDTO.getName();
-    }
-
-    public Long getId() {
-        return userDTO.getUserId();
-    }
-}
