--- src/main/java/oauth/test/oauth2jwt/dto/OAuth2Response.java
@@ -1,12 +1,0 @@
-package oauth.test.oauth2jwt.dto;
-
-public interface OAuth2Response {
-    //제공자 (Ex. naver, google, ...)
-    String getProvider();
-    //제공자에서 발급해주는 아이디(번호)
-    String getProviderId();
-    //이메일
-    String getEmail();
-    //사용자 실명 (설정한 이름)
-    String getName();
-}
