--- src/main/java/oauth/test/oauth2jwt/jwt/CustomAuthenticationEntryPoint.java
@@ -1,21 +1,0 @@
-package oauth.test.oauth2jwt.jwt;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Component
-public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
-
-    @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response,
-                         AuthenticationException authException) throws IOException {
-        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED); // 401
-        response.setContentType("application/json;charset=UTF-8");
-        response.getWriter().write("{\"error\": \"Unauthorized or expired token\"}");
-    }
-}
