--- src/main/java/oauth/test/oauth2jwt/oauth2/CustomSuccessHandler.java
@@ -1,60 +1,0 @@
-package oauth.test.oauth2jwt.oauth2;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import oauth.test.oauth2jwt.dto.CustomOAuth2User;
-import oauth.test.oauth2jwt.jwt.JWTUtil;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-import java.util.Collection;
-import java.util.Iterator;
-
-@Component
-@RequiredArgsConstructor
-public class CustomSuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
-
-    private final JWTUtil jwtUtil;
-
-    @Value("${client.url}")
-    String clientUrl;
-
-    @Override
-    public void onAuthenticationSuccess(HttpServletRequest request,
-                                        HttpServletResponse response,
-                                        Authentication authentication
-    ) throws IOException, ServletException {
-
-        //OAuth2User
-        CustomOAuth2User customUserDetails = (CustomOAuth2User) authentication.getPrincipal();
-
-        Long userId = customUserDetails.getId();
-
-        Collection<? extends GrantedAuthority> authorities = authentication.getAuthorities();
-        Iterator<? extends GrantedAuthority> iterator = authorities.iterator();
-        GrantedAuthority auth = iterator.next();
-        String role = auth.getAuthority();
-
-        String token = jwtUtil.createJwt(userId, role, 60 * 60 * 1000 * 3L); // 3시간
-
-        // Cookie에 담아 전달
-        response.addCookie(createCookie("Authorization", token));
-        response.sendRedirect(clientUrl); // 클라이언트 url
-    }
-
-    private Cookie createCookie(String key, String value) {
-        Cookie cookie = new Cookie(key, value);
-        cookie.setMaxAge(60*60*60);
-        //cookie.setSecure(true); //https만 가능하게 설정
-        cookie.setPath("/");
-//        cookie.setHttpOnly(true); //js가 cookie 못가져가도록 설정
-        return cookie;
-    }
-}
