--- src/main/java/oauth/test/oauth2jwt/repository/UserRepository.java
@@ -1,8 +1,0 @@
-package oauth.test.oauth2jwt.repository;
-
-import oauth.test.oauth2jwt.entity.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-    UserEntity findByOauthCode(String oauthCode);
-}
