+++ src/main/resources/application-dev.properties
@@ -1,3 +1,6 @@
+# logging
+logging.level.root=debug
+
 # mysql
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.url=jdbc:mysql://127.0.0.1:3306/konnect?useSSL=false&useUnicode=true&serverTimezone=Asia/Seoul&allowPublicKeyRetrieval=true&createDatabaseIfNotExist=true
