+++ src/main/resources/application-prod.properties
@@ -1,3 +1,5 @@
+# logging
+logging.level.root=info
 # mysql
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.url=jdbc:mysql://amazone:3306/ssafy?useSSL=false&useUnicode=true&serverTimezone=Asia/Seoul&allowPublicKeyRetrieval=true
