+++ src/main/resources/application.properties
@@ -1,9 +1,9 @@
-spring.application.name=OAuth2JWT
-
-logging.level.root=info
+spring.application.name=Konnect
 
 # ?? ?? ???? ??
 spring.profiles.active=dev
+
+spring.jpa.open-in-view=false
 
 # properties
 spring.config.import=classpath:application-secret.properties
