+++ src/test/java/com/konnect/KonnectApplicationTests.java
@@ -1,0 +1,13 @@
+package com.konnect;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class KonnectApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
