--- src/test/java/oauth/test/oauth2jwt/OAuth2JwtApplicationTests.java
@@ -1,13 +1,0 @@
-package oauth.test.oauth2jwt;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class OAuth2JwtApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
